Added Retake and Accept Buttons to Photo
