+++ src/CameraModule/ImagePreview.css
@@ -1,0 +1,40 @@
+.demo-image-preview {
+  position: relative;
+  text-align: center;
+}
+
+.demo-image-preview>img {
+  width: 768px;
+}
+
+@media(max-width:768px) {
+  .demo-image-preview>img {
+    width: 100%;
+  }
+}
+
+/* fullscreen enable by props */
+
+.demo-image-preview-fullscreen>img {
+  width: 100vw;
+  height: 100vh;
+}
+
+.IconRetake {
+  color: red;
+  position: absolute;
+  padding-left: 15px;
+  padding-top: 15px;
+  z-index: 2;
+}
+
+.IconAccept {
+  width: 100%;
+  padding-top: 15px;
+  padding-right: 15px;
+  color: green;
+  position: absolute;
+  display: flex;
+  justify-content: flex-end;
+  z-index: 1;
+}
